style: 전체 페이지 배경색 추가, 레이아웃 수정
