+++ src/app/(main)/layout.tsx
@@ -5,11 +5,11 @@ import Footer from '@/components/common/Footer';
 const MainLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <>
-      <div className="relative min-h-screen flex flex-col">
-        <div className="flex-none">
-          <Header />
-        </div>
-        <div className="flex-grow flex flex-col justify-center items-center pt-15">{children}</div>
+      <div className="relative min-h-screen">
+        <Header />
+        <main className="pt-15 bg-default">
+          <div className="max-w-[1440px] px-10 py-10 my-0 mx-auto min-h-screen">{children}</div>
+        </main>
         <Footer />
       </div>
     </>
