+++ src/app/(main)/login/page.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 
 const LoginPage = () => {
   return (
-    <div className="min-h-screen w-full bg-gray-100 flex flex-col justify-center">
+    <div className="w-full flex flex-col justify-center pt-[183px]">
       <div className="flex flex-col">
         <div className="flex flex-col items-center mb-20">
           <h2 className="login_title text-2xl font-bold mb-4">Health Protect me</h2>
