+++ src/components/Contents/Posting/Category/Categories.tsx
@@ -15,25 +15,23 @@ const CategoryMain = ({ categories, selectedCategories, onSelectCategory }: Cate
   }, [selectedCategories]);
 
   return (
-    <div className="mb-8">
-      <div>
-        {categories.map((category) => (
-          <button
-            key={category.id}
-            onClick={() => {
-              onSelectCategory(category.id);
-              setShowError(false);
-            }}
-            className={`px-4 py-2 ml-2 border-gray300 rounded-xl w-[72px] ${
-              selectedCategories.includes(category.id)
-                ? 'bg-pramary600 text-white'
-                : 'bg-white text-gray900 border-spacing-1 border border-solid'
-            } hover:bg-pramary600 hover:text-white transition-colors duration-300`}
-          >
-            {category.name}
-          </button>
-        ))}
-      </div>
+    <div className="mb-6">
+      {categories.map((category) => (
+        <button
+          key={category.id}
+          onClick={() => {
+            onSelectCategory(category.id);
+            setShowError(false);
+          }}
+          className={`px-4 py-2 mr-2 border-gray300 rounded-xl w-[72px] ${
+            selectedCategories.includes(category.id)
+              ? 'bg-pramary600 text-white'
+              : 'bg-white text-gray900 border-spacing-1 border border-solid'
+          } hover:bg-pramary600 hover:text-white transition-colors duration-300`}
+        >
+          {category.name}
+        </button>
+      ))}
     </div>
   );
 };
