+++ src/components/Contents/Posting/PostingMain.tsx
@@ -1,11 +1,7 @@
 import TextareatPage from './Textarea/TextareatPage';
 
 const PostingMain = () => {
-  return (
-    <div>
-      <TextareatPage />
-    </div>
-  );
+  return <TextareatPage />;
 };
 
 export default PostingMain;
