+++ src/components/MyPage/MyProfile.tsx
@@ -3,8 +3,8 @@ import PostsSection from '@/components/MyPage/ProfilePage/PostsSection';
 
 const MyProfile = () => {
   return (
-    <main className="w-[1440px] h-[858px] flex justify-between bg-white p-10">
-      <div className="max-w-full mx-10 flex flex-col md:flex-row">
+    <div className="h-[858px] flex justify-between">
+      <div className="max-w-full flex flex-col md:flex-row">
         <div className="flex flex-col items-start mr-10">
           <Profile />
         </div>
@@ -12,7 +12,7 @@ const MyProfile = () => {
           <PostsSection />
         </div>
       </div>
-    </main>
+    </div>
   );
 };
