+++ src/components/MyPage/profileEdit/ProfileEdit.tsx
@@ -216,7 +216,7 @@ const ProfileEdit = ({
   };
 
   return (
-    <section className="w-[1360px] max-w-md mx-auto mt-10">
+    <section className="w-[1360px] max-w-md mx-auto">
       <h1 className="w-[400px] h-8 mb-4 mx-4 text-2xl font-bold">프로필 수정</h1>
       <div className="flex flex-col items-center text-center mb-8 w-full px-4">
         <div className="flex justify-between items-center w-full mb-10">
