+++ tailwind.config.ts
@@ -92,11 +92,15 @@ const config = {
         pramary100: '#FFF6F2',
         pramary500: '#FF7A85',
         pramary600: '#F5637C',
-        btnClose: 'rgba(39, 40, 42, 0.50)'
+        btnClose: 'rgba(39, 40, 42, 0.50)',
+        default: '#F8FAF8'
       },
       boxShadow: {
         'main-btn': '0px 8px 16px 0px rgba(255, 122, 133, 0.40)',
         'main-btn-hover': '2px 4px 4px 0px rgba(39, 40, 42, 0.25), 0px 8px 16px 0px rgba(255, 122, 133, 0.40)'
+      },
+      height: {
+        'main-height': 'calc(100vh - 57px)'
       }
     }
   },
